separate out Config from Utils
