+++ src/main/scala/backends/Backend.scala
@@ -1,13 +1,14 @@
 package fuselang.backend
 
-import fuselang.CompilerError.BackendError
+import fuselang._
+import CompilerError.BackendError
 
 /**
  * Abstract definition of a Fuse backend.
  */
 trait Backend {
 
-  def emit(p: fuselang.Syntax.Prog, c: fuselang.Utils.Config): String = {
+  def emit(p: Syntax.Prog, c: Configuration.Config): String = {
     if (c.header && (canGenerateHeader == false)) {
       throw BackendError(s"Backend $this does not support header generation.")
     }
@@ -19,7 +20,7 @@ trait Backend {
    * program. Assumes that typechecking pass has been done on the AST and
    * might use the type decorations added to the nodes.
    */
-  def emitProg(p: fuselang.Syntax.Prog, c: fuselang.Utils.Config): String
+  def emitProg(p: Syntax.Prog, c: Configuration.Config): String
 
   /**
    * True if the backend can generate header files.
